# Remove unnecessary ResetTimer() for symmetry