@fredrb 2022-12-09
@vikin91 No significant difference in the benchmarks:

```
% benchstat ./result.old.txt ./result.new.txt
name                      old time/op    new time/op    delta
Find/K=0-N=10^2             7.77µs ± 0%    7.87µs ± 0%   ~     (p=1.000 n=1+1)
Find/K=0-N=10^3              124µs ± 0%     122µs ± 0%   ~     (p=1.000 n=1+1)
Find/K=0-N=10^4             6.21ms ± 0%    2.94ms ± 0%   ~     (p=1.000 n=1+1)
Find/K=0-N=10^5             59.5ms ± 0%    61.7ms ± 0%   ~     (p=1.000 n=1+1)
Find/K=0-N=10^6              628ms ± 0%     741ms ± 0%   ~     (p=1.000 n=1+1)
Find/K=1-N=10^2             86.4µs ± 0%   104.9µs ± 0%   ~     (p=1.000 n=1+1)
Find/K=1-N=10^3             1.10ms ± 0%    1.35ms ± 0%   ~     (p=1.000 n=1+1)
Find/K=1-N=10^4             19.8ms ± 0%    23.9ms ± 0%   ~     (p=1.000 n=1+1)
Find/K=1-N=10^5              279ms ± 0%     308ms ± 0%   ~     (p=1.000 n=1+1)
Find/K=1-N=10^6              4.16s ± 0%     4.36s ± 0%   ~     (p=1.000 n=1+1)
Find/K=2-N=10^2             91.8µs ± 0%    98.1µs ± 0%   ~     (p=1.000 n=1+1)
Find/K=2-N=10^3             1.25ms ± 0%    1.25ms ± 0%   ~     (p=1.000 n=1+1)
Find/K=2-N=10^4             21.7ms ± 0%    21.8ms ± 0%   ~     (p=1.000 n=1+1)
Find/K=2-N=10^5              305ms ± 0%     306ms ± 0%   ~     (p=1.000 n=1+1)
Find/K=2-N=10^6              4.20s ± 0%     3.78s ± 0%   ~     (p=1.000 n=1+1)
Find/K=3-N=10^2              123µs ± 0%     102µs ± 0%   ~     (p=1.000 n=1+1)
Find/K=3-N=10^3             1.32ms ± 0%    1.32ms ± 0%   ~     (p=1.000 n=1+1)
Find/K=3-N=10^4             22.0ms ± 0%    22.2ms ± 0%   ~     (p=1.000 n=1+1)
Find/K=3-N=10^5              318ms ± 0%     304ms ± 0%   ~     (p=1.000 n=1+1)
Find/K=3-N=10^6              4.31s ± 0%     4.16s ± 0%   ~     (p=1.000 n=1+1)
Find/K=4-N=10^2              101µs ± 0%      94µs ± 0%   ~     (p=1.000 n=1+1)
Find/K=4-N=10^3             1.46ms ± 0%    1.20ms ± 0%   ~     (p=1.000 n=1+1)
Find/K=4-N=10^4             31.4ms ± 0%    19.7ms ± 0%   ~     (p=1.000 n=1+1)
Find/K=4-N=10^5              287ms ± 0%     285ms ± 0%   ~     (p=1.000 n=1+1)
Find/K=4-N=10^6              3.64s ± 0%     3.90s ± 0%   ~     (p=1.000 n=1+1)
Find/K=5-N=10^2             96.2µs ± 0%    87.8µs ± 0%   ~     (p=1.000 n=1+1)
Find/K=5-N=10^3             1.27ms ± 0%    1.15ms ± 0%   ~     (p=1.000 n=1+1)
Find/K=5-N=10^4             21.9ms ± 0%    20.8ms ± 0%   ~     (p=1.000 n=1+1)
Find/K=5-N=10^5              311ms ± 0%     307ms ± 0%   ~     (p=1.000 n=1+1)
Find/K=5-N=10^6              4.15s ± 0%     4.21s ± 0%   ~     (p=1.000 n=1+1)
Find/K=6-N=10^2             90.6µs ± 0%   114.7µs ± 0%   ~     (p=1.000 n=1+1)
Find/K=6-N=10^3             1.24ms ± 0%    1.31ms ± 0%   ~     (p=1.000 n=1+1)
Find/K=6-N=10^4             21.5ms ± 0%    21.9ms ± 0%   ~     (p=1.000 n=1+1)
Find/K=6-N=10^5              314ms ± 0%     305ms ± 0%   ~     (p=1.000 n=1+1)
Find/K=6-N=10^6              3.46s ± 0%     4.38s ± 0%   ~     (p=1.000 n=1+1)
Find/K=7-N=10^2             93.1µs ± 0%   100.3µs ± 0%   ~     (p=1.000 n=1+1)
Find/K=7-N=10^3             1.33ms ± 0%    1.35ms ± 0%   ~     (p=1.000 n=1+1)
Find/K=7-N=10^4             21.3ms ± 0%    23.0ms ± 0%   ~     (p=1.000 n=1+1)
Find/K=7-N=10^5              318ms ± 0%     315ms ± 0%   ~     (p=1.000 n=1+1)
Find/K=7-N=10^6              4.29s ± 0%     4.86s ± 0%   ~     (p=1.000 n=1+1)
Find/K=8-N=10^2              111µs ± 0%      89µs ± 0%   ~     (p=1.000 n=1+1)
Find/K=8-N=10^3             1.69ms ± 0%    1.30ms ± 0%   ~     (p=1.000 n=1+1)
Find/K=8-N=10^4             21.8ms ± 0%    21.3ms ± 0%   ~     (p=1.000 n=1+1)
Find/K=8-N=10^5              258ms ± 0%     296ms ± 0%   ~     (p=1.000 n=1+1)
Find/K=8-N=10^6              4.11s ± 0%     3.93s ± 0%   ~     (p=1.000 n=1+1)
Find/K=9-N=10^2             90.0µs ± 0%    86.4µs ± 0%   ~     (p=1.000 n=1+1)
Find/K=9-N=10^3             1.17ms ± 0%    1.15ms ± 0%   ~     (p=1.000 n=1+1)
Find/K=9-N=10^4             19.2ms ± 0%    19.2ms ± 0%   ~     (p=1.000 n=1+1)
Find/K=9-N=10^5              291ms ± 0%     287ms ± 0%   ~     (p=1.000 n=1+1)
Find/K=9-N=10^6              3.93s ± 0%     3.92s ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=0-N=10^2    1.43µs ± 0%    1.45µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=0-N=10^3    1.43µs ± 0%    1.44µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=0-N=10^4    1.56µs ± 0%    1.51µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=0-N=10^5    1.88µs ± 0%    1.83µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=0-N=10^6    2.40µs ± 0%    2.33µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=1-N=10^2    1.44µs ± 0%    1.51µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=1-N=10^3    1.42µs ± 0%    1.46µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=1-N=10^4    1.49µs ± 0%    1.55µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=1-N=10^5    1.77µs ± 0%    1.92µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=1-N=10^6    2.63µs ± 0%    2.58µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=2-N=10^2    1.44µs ± 0%    1.91µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=2-N=10^3    1.43µs ± 0%    1.48µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=2-N=10^4    1.50µs ± 0%    1.51µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=2-N=10^5    1.74µs ± 0%    1.88µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=2-N=10^6    1.76µs ± 0%    2.56µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=3-N=10^2    1.44µs ± 0%    1.50µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=3-N=10^3    1.60µs ± 0%    1.43µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=3-N=10^4    1.90µs ± 0%    1.48µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=3-N=10^5    1.91µs ± 0%    1.85µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=3-N=10^6    1.79µs ± 0%    2.44µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=4-N=10^2    1.41µs ± 0%    1.41µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=4-N=10^3    1.42µs ± 0%    1.44µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=4-N=10^4    1.47µs ± 0%    1.46µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=4-N=10^5    1.82µs ± 0%    1.89µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=4-N=10^6    1.54µs ± 0%    2.42µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=5-N=10^2    1.44µs ± 0%    1.43µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=5-N=10^3    1.48µs ± 0%    1.40µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=5-N=10^4    1.50µs ± 0%    1.50µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=5-N=10^5    1.84µs ± 0%    1.81µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=5-N=10^6    2.56µs ± 0%    1.18µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=6-N=10^2    1.45µs ± 0%    1.36µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=6-N=10^3    1.42µs ± 0%    1.34µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=6-N=10^4    1.50µs ± 0%    1.40µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=6-N=10^5    1.84µs ± 0%    1.90µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=6-N=10^6    2.72µs ± 0%    2.43µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=7-N=10^2    1.42µs ± 0%    1.45µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=7-N=10^3    1.41µs ± 0%    1.47µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=7-N=10^4    1.49µs ± 0%    1.53µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=7-N=10^5    1.81µs ± 0%    1.90µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=7-N=10^6    2.46µs ± 0%    1.43µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=8-N=10^2    1.39µs ± 0%    1.40µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=8-N=10^3    1.39µs ± 0%    1.39µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=8-N=10^4    1.48µs ± 0%    1.45µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=8-N=10^5    1.73µs ± 0%    1.88µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=8-N=10^6    2.60µs ± 0%    2.58µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=9-N=10^2    1.42µs ± 0%    1.46µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=9-N=10^3    1.40µs ± 0%    1.45µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=9-N=10^4    1.46µs ± 0%    1.55µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=9-N=10^5    1.80µs ± 0%    1.87µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Update/L=9-N=10^6    2.36µs ± 0%    1.29µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=0-N=10^2       4.08µs ± 0%    3.52µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=0-N=10^3       3.85µs ± 0%    3.62µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=0-N=10^4       4.21µs ± 0%    4.24µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=0-N=10^5       3.81µs ± 0%    4.06µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=0-N=10^6       3.57µs ± 0%    3.49µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=1-N=10^2       3.69µs ± 0%    3.83µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=1-N=10^3       3.69µs ± 0%    3.80µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=1-N=10^4       4.49µs ± 0%    4.19µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=1-N=10^5       4.61µs ± 0%    4.36µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=1-N=10^6       2.75µs ± 0%    3.50µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=2-N=10^2       4.39µs ± 0%    3.55µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=2-N=10^3       3.90µs ± 0%    3.53µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=2-N=10^4       5.25µs ± 0%    4.39µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=2-N=10^5       5.06µs ± 0%    4.15µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=2-N=10^6       3.29µs ± 0%    3.64µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=3-N=10^2       3.87µs ± 0%    3.51µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=3-N=10^3       4.42µs ± 0%    3.58µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=3-N=10^4       4.62µs ± 0%    4.07µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=3-N=10^5       5.13µs ± 0%    4.48µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=3-N=10^6       2.61µs ± 0%    3.43µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=4-N=10^2       4.15µs ± 0%    3.78µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=4-N=10^3       4.49µs ± 0%    3.60µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=4-N=10^4       4.97µs ± 0%    4.23µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=4-N=10^5       4.69µs ± 0%    4.19µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=4-N=10^6       3.15µs ± 0%    3.19µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=5-N=10^2       3.99µs ± 0%    3.82µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=5-N=10^3       4.25µs ± 0%    3.62µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=5-N=10^4       4.97µs ± 0%    4.32µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=5-N=10^5       4.68µs ± 0%    4.22µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=5-N=10^6       2.62µs ± 0%    3.63µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=6-N=10^2       4.38µs ± 0%    3.68µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=6-N=10^3       4.60µs ± 0%    3.61µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=6-N=10^4       5.03µs ± 0%    4.16µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=6-N=10^5       4.95µs ± 0%    4.29µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=6-N=10^6       4.33µs ± 0%    3.36µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=7-N=10^2       5.24µs ± 0%    3.62µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=7-N=10^3       4.38µs ± 0%    3.79µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=7-N=10^4       4.47µs ± 0%    3.94µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=7-N=10^5       4.64µs ± 0%    3.88µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=7-N=10^6       2.79µs ± 0%    3.05µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=8-N=10^2       4.33µs ± 0%    3.80µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=8-N=10^3       4.80µs ± 0%    3.71µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=8-N=10^4       4.55µs ± 0%    4.10µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=8-N=10^5       4.71µs ± 0%    4.27µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=8-N=10^6       2.57µs ± 0%    3.30µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=9-N=10^2       4.21µs ± 0%    3.47µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=9-N=10^3       4.31µs ± 0%    3.57µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=9-N=10^4       4.27µs ± 0%    4.21µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=9-N=10^5       4.80µs ± 0%    4.14µs ± 0%   ~     (p=1.000 n=1+1)
Upsert_Add/L=9-N=10^6       3.28µs ± 0%    3.27µs ± 0%   ~     (p=1.000 n=1+1)
```