@fredrb 2022-12-09
@vikin91 there seems to be some difference, but seems a bit random. Since I've ran them in my machine, other processes could have influenced the result. 

From the code change perspective I don't think it's a problem to merge this. WDYT?

Ran using `-count=5` to compare. 

```
% benchstat ./result.old.txt ./result.new.txt
name                      old time/op    new time/op    delta
Find/K=0-N=10^2             7.18µs ± 1%    7.38µs ± 2%   +2.88%  (p=0.008 n=5+5)
Find/K=0-N=10^4             1.99ms ± 9%    3.41ms ±21%  +71.25%  (p=0.008 n=5+5)
Find/K=0-N=10^6              700ms ± 5%     725ms ± 3%     ~     (p=0.151 n=5+5)
Find/K=2-N=10^2             86.7µs ± 8%    88.8µs ± 7%     ~     (p=0.222 n=5+5)
Find/K=2-N=10^4             20.7ms ± 4%    21.3ms ± 4%     ~     (p=0.222 n=5+5)
Find/K=2-N=10^6              3.97s ± 7%     3.96s ± 6%     ~     (p=1.000 n=5+5)
Find/K=4-N=10^2             86.1µs ± 4%    90.7µs ± 9%     ~     (p=0.095 n=5+5)
Find/K=4-N=10^4             20.1ms ± 6%    21.9ms ± 3%   +8.98%  (p=0.008 n=5+5)
Find/K=4-N=10^6              3.87s ± 5%     4.10s ± 2%   +5.89%  (p=0.032 n=5+5)
Find/K=9-N=10^2             83.0µs ± 2%    89.5µs ± 3%   +7.90%  (p=0.008 n=5+5)
Find/K=9-N=10^4             21.4ms ± 9%    21.1ms ± 2%     ~     (p=0.690 n=5+5)
Find/K=9-N=10^6              4.04s ± 7%     4.09s ± 0%     ~     (p=0.151 n=5+5)
Upsert_Update/L=0-N=10^2    1.52µs ±13%    1.51µs ± 4%     ~     (p=1.000 n=5+5)
Upsert_Update/L=0-N=10^4    1.51µs ± 2%    1.59µs ± 2%   +5.13%  (p=0.008 n=5+5)
Upsert_Update/L=0-N=10^6    2.52µs ± 9%    2.63µs ± 6%     ~     (p=0.841 n=5+5)
Upsert_Update/L=2-N=10^2    1.51µs ± 6%    1.54µs ± 4%     ~     (p=0.421 n=5+5)
Upsert_Update/L=2-N=10^4    1.70µs ± 3%    1.72µs ± 3%     ~     (p=0.421 n=5+5)
Upsert_Update/L=2-N=10^6    2.47µs ± 8%    2.49µs ± 6%     ~     (p=1.000 n=5+5)
Upsert_Update/L=4-N=10^2    1.58µs ±16%    1.56µs ± 6%     ~     (p=0.841 n=5+5)
Upsert_Update/L=4-N=10^4    1.68µs ± 9%    1.62µs ± 7%     ~     (p=0.310 n=5+5)
Upsert_Update/L=4-N=10^6    2.53µs ±10%    2.51µs ±13%     ~     (p=0.690 n=5+5)
Upsert_Update/L=9-N=10^2    1.51µs ± 2%    1.56µs ± 7%     ~     (p=0.278 n=5+5)
Upsert_Update/L=9-N=10^4    1.60µs ± 5%    1.73µs ± 4%   +7.78%  (p=0.016 n=5+5)
Upsert_Update/L=9-N=10^6    2.35µs ± 6%    1.50µs ±41%  -36.23%  (p=0.008 n=5+5)
Upsert_Add/L=0-N=10^2       4.43µs ±14%    5.08µs ±26%     ~     (p=0.310 n=5+5)
Upsert_Add/L=0-N=10^4       4.92µs ±21%    4.94µs ±26%     ~     (p=1.000 n=5+5)
Upsert_Add/L=0-N=10^6       3.53µs ±50%    2.82µs ±25%     ~     (p=0.151 n=5+5)
Upsert_Add/L=2-N=10^2       5.02µs ±37%    5.21µs ±37%     ~     (p=0.841 n=5+5)
Upsert_Add/L=2-N=10^4       5.25µs ±24%    5.49µs ±13%     ~     (p=0.841 n=5+5)
Upsert_Add/L=2-N=10^6       2.63µs ±13%    3.56µs ±42%  +35.46%  (p=0.032 n=5+5)
Upsert_Add/L=4-N=10^2       5.37µs ±28%    5.04µs ±18%     ~     (p=0.548 n=5+5)
Upsert_Add/L=4-N=10^4       5.00µs ±20%    5.18µs ±20%     ~     (p=0.690 n=5+5)
Upsert_Add/L=4-N=10^6       3.32µs ±40%    2.75µs ±21%     ~     (p=0.095 n=5+5)
Upsert_Add/L=9-N=10^2       4.67µs ± 9%    5.04µs ±23%     ~     (p=0.421 n=5+5)
Upsert_Add/L=9-N=10^4       5.12µs ±27%    5.42µs ±27%     ~     (p=1.000 n=5+5)
Upsert_Add/L=9-N=10^6       3.23µs ±36%    3.28µs ±30%     ~     (p=0.690 n=5+5)
```